MapMessage & MapMessages utils
